requirements.txt
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
feedparser==6.0.10
google-api-python-client==2.80.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
httplib2==0.21.0
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.14
